+++ src/pyramid_bimt/scripts/populate.py
@@ -224,7 +224,7 @@ def add_demo_auditlog_entries():
     with transaction.manager:
         read = AuditLogEntry(
             user=User.by_email('staff@bar.com'),
-            event_type_id=1,
+            event_type_id=2,
             comment='read entry',
             read=True,
         )
@@ -232,7 +232,7 @@ def add_demo_auditlog_entries():
 
         unread = AuditLogEntry(
             user=User.by_email('one@bar.com'),
-            event_type_id=1,
+            event_type_id=2,
             comment='unread entry',
             read=False,
         )