+++ src/pyramid_bimt/scripts/sanitycheck_email.py
@@ -15,23 +15,23 @@ import sys
 logger = logging.getLogger(__name__)
 
 
-def send_email(settings, request):
-    warnings = run_all_checks(request.registry)
+def send_email(settings, request, verbose=False):
+    warnings = run_all_checks(request)
     if warnings:
         subject = '{} sanity check found {} warnings'.format(
             settings['bimt.app_title'], len(warnings))
     else:
         subject = '{} sanity check OK'.format(settings['bimt.app_title'])
-
-    mailer = get_mailer(request)
-    mailer.send_immediately(Message(
-        recipients=['maintenance@niteoweb.com', ],
-        subject=subject,
-        html=render(
-            'pyramid_bimt:templates/sanitycheck_email.pt',
-            {'warnings': warnings},
-        ),
-    ))
+    if warnings or verbose:
+        mailer = get_mailer(request)
+        mailer.send_immediately(Message(
+            recipients=['maintenance@niteoweb.com', ],
+            subject=subject,
+            html=render(
+                'pyramid_bimt:templates/sanitycheck_email.pt',
+                {'warnings': warnings},
+            ),
+        ))
 
 
 def main(argv=sys.argv):
@@ -42,13 +42,19 @@ def main(argv=sys.argv):
     parser.add_argument(
         'config', type=str, metavar='<config>',
         help='Pyramid application configuration file.')
+    parser.add_argument(
+        '-v', '--verbose',
+        help='Send email regardless of sanity check result.',
+        action='store_true',
+    )
 
     env = bootstrap(parser.parse_args().config)
     setup_logging(parser.parse_args().config)
+    verbose = parser.parse_args().verbose
 
     settings = env['registry'].settings
     request = env['request']
-    send_email(settings, request)
+    send_email(settings, request, verbose=verbose)
 
     env['closer']()
     logger.info('Sanity check email sent successfully.')